Updated folder structure and separated by database, factory, manager, model.
