+++ src/main/java/com/mycompany/librarymanagementsystem/BookOperations.java
@@ -4,6 +4,8 @@
  */
 package com.mycompany.librarymanagementsystem;
 
+import com.mycompany.librarymanagementsystem.model.Book;
+
 /**
  *
  * @author francescosciabbarrasi
