+++ src/main/java/com/mycompany/librarymanagementsystem/manager/UserManager.java
@@ -2,7 +2,10 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package com.mycompany.librarymanagementsystem;
+package com.mycompany.librarymanagementsystem.manager;
+import com.mycompany.librarymanagementsystem.db.DatabaseConnection;
+import com.mycompany.librarymanagementsystem.model.User;
+import com.mycompany.librarymanagementsystem.UserOperations;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
