+++ src/main/java/com/mycompany/librarymanagementsystem/model/User.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package com.mycompany.librarymanagementsystem;
+package com.mycompany.librarymanagementsystem.model;
 
 /**
  *
